Add files via upload
I started off with 2 circle SOPS and changed the dimensions of each one. I connected these SOPS with a copy SOP. I created my Geometry network, and I assigned a line material to it. I used a rgb key TOP to change the background to black. I am proud of this work because I was struggling with geometry networks in class, but after this piece I got the hang of it. I would like to develop my geometry networks in the future by making it audio reactive , as I think that would look really good.