Add files via upload
week 8 -feedback/transform - touch designer
for this code, I started with a circle TOP and turned the dimensions down to make it a triangle. I connected a transform and feedback TOP, using absTime.seconds I animated the rotation of the triangle. I changed the colours in the circle TOP to green and white. I am proud of this outcome because it was my first creative code that I did entirely by-myself in touch designer. It took alot of trial and error to get my desired outcome such as the right rotation speed, but  it worked out in the end. If i could make improvements to this code in the future I would make improvements with the aesthetics, maybe adding a lookup or bloom TOP.

also this file has 2 peices of code in it, I am describing the first green circle code.